- Reduced lodash bundle size. - Removed firestore import on _app - Deleted unwanted admin list page - Added Bundle Analyzer - Added Transpile Module support to transform lodash imports
