+++ .babelrc
@@ -1,3 +1,14 @@
 {
-    "presets": ["next/babel"]
+    "presets": ["next/babel"],
+    "plugins": [
+        [
+            "babel-plugin-transform-imports",
+            {
+                "lodash": {
+                    "transform": "lodash/${member}",
+                    "preventFullImport": true
+                }
+            }
+        ]
+    ]
 }
