+++ package.json
@@ -11,7 +11,10 @@
         "format": "prettier --write .",
         "lint": "eslint . --ext ts --ext tsx --ext js",
         "test": "jest",
-        "test-all": "npm run lint && npm run type-check && npm run test"
+        "test-all": "npm run lint && npm run type-check && npm run test",
+        "analyze": "cross-env ANALYZE=true next build",
+        "analyze:server": "cross-env BUNDLE_ANALYZE=server next build",
+        "analyze:browser": "cross-env BUNDLE_ANALYZE=browser next build"
     },
     "husky": {
         "hooks": {
@@ -30,6 +33,8 @@
         "@chakra-ui/react": "^1.6.5",
         "@emotion/react": "^11.4.0",
         "@emotion/styled": "^11.3.0",
+        "@next/bundle-analyzer": "^11.1.0",
+        "cross-env": "^7.0.3",
         "eslint-config-next": "^11.0.1",
         "final-form": "^4.20.2",
         "firebase": "^8.8.0",
@@ -39,6 +44,7 @@
         "next": "latest",
         "next-firebase-auth": "^0.13.2",
         "next-seo": "^4.26.0",
+        "next-transpile-modules": "^8.0.0",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-final-form": "^6.5.3",
@@ -54,6 +60,7 @@
         "@typescript-eslint/eslint-plugin": "^4.14.2",
         "@typescript-eslint/parser": "^4.14.2",
         "babel-jest": "^27.0.2",
+        "babel-plugin-transform-imports": "^2.0.0",
         "eslint": "^7.19.0",
         "eslint-config-prettier": "^7.2.0",
         "eslint-plugin-react": "^7.19.0",
